rename vec to simdvec
