+++ libs/native/src/main/java/module-info.java
@@ -19,8 +19,8 @@ module org.elasticsearch.nativeaccess {
         to
             org.elasticsearch.nativeaccess.jna,
             org.elasticsearch.server,
-            org.elasticsearch.systemd,
-            org.elasticsearch.vec;
+            org.elasticsearch.simdvec,
+            org.elasticsearch.systemd;
     // allows jna to implement a library provider, and ProviderLocator to load it
     exports org.elasticsearch.nativeaccess.lib to org.elasticsearch.nativeaccess.jna, org.elasticsearch.base;
