+++ libs/simdvec/build.gradle
