+++ libs/simdvec/licenses/lucene-core-LICENSE.txt
