+++ libs/simdvec/licenses/lucene-core-NOTICE.txt
