+++ libs/simdvec/native/build.gradle
