+++ libs/simdvec/native/publish_vec_binaries.sh
