+++ libs/simdvec/native/settings.gradle
