+++ libs/simdvec/native/src/vec/c/aarch64/vec.c
