+++ libs/simdvec/native/src/vec/c/amd64/vec.c
