--- libs/vec/src/main21/java/org/elasticsearch/vec/internal/Similarities.java
@@ -1,53 +1,0 @@
-/*
- * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
- * or more contributor license agreements. Licensed under the Elastic License
- * 2.0 and the Server Side Public License, v 1; you may not use this file except
- * in compliance with, at your election, the Elastic License 2.0 or the Server
- * Side Public License, v 1.
- */
-
-package org.elasticsearch.vec.internal;
-
-import org.elasticsearch.nativeaccess.NativeAccess;
-import org.elasticsearch.nativeaccess.VectorSimilarityFunctions;
-
-import java.lang.foreign.MemorySegment;
-import java.lang.invoke.MethodHandle;
-
-public class Similarities {
-
-    static final VectorSimilarityFunctions DISTANCE_FUNCS = NativeAccess.instance()
-        .getVectorSimilarityFunctions()
-        .orElseThrow(AssertionError::new);
-
-    static final MethodHandle DOT_PRODUCT_7U = DISTANCE_FUNCS.dotProductHandle7u();
-    static final MethodHandle SQUARE_DISTANCE_7U = DISTANCE_FUNCS.squareDistanceHandle7u();
-
-    static int dotProduct7u(MemorySegment a, MemorySegment b, int length) {
-        try {
-            return (int) DOT_PRODUCT_7U.invokeExact(a, b, length);
-        } catch (Throwable e) {
-            if (e instanceof Error err) {
-                throw err;
-            } else if (e instanceof RuntimeException re) {
-                throw re;
-            } else {
-                throw new RuntimeException(e);
-            }
-        }
-    }
-
-    static int squareDistance7u(MemorySegment a, MemorySegment b, int length) {
-        try {
-            return (int) SQUARE_DISTANCE_7U.invokeExact(a, b, length);
-        } catch (Throwable e) {
-            if (e instanceof Error err) {
-                throw err;
-            } else if (e instanceof RuntimeException re) {
-                throw re;
-            } else {
-                throw new RuntimeException(e);
-            }
-        }
-    }
-}
