+++ server/src/main/java/module-info.java
@@ -32,7 +32,7 @@ module org.elasticsearch.server {
     requires org.elasticsearch.plugin.analysis;
     requires org.elasticsearch.grok;
     requires org.elasticsearch.tdigest;
-    requires org.elasticsearch.vec;
+    requires org.elasticsearch.simdvec;
 
     requires com.sun.jna;
     requires hppc;
